+++ CHANGELOG.md
@@ -15,9 +15,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add primitives_mapping property: returns a dictionary containing the correspondence between the parametric and 3D boundaries of the faces.
 - grid_points: returns a grid of points inside the surface2d of the face.
 
+#### surfaces.py
+- CylindricalSurface3D: parametric_points_to_3d
+- ToroidalSurface3D: parametric_points_to_3d
+- SphericalSurface3D: parametric_points_to_3d
+- ConicalSurface3D: parametric_points_to_3d
+- ExtrusionSurface3D: parametric_points_to_3d
+- RevolutionSurface3D: parametric_points_to_3d
+- Plane3D: parametric_points_to_3d
+- BSplineSurface3D: parametric_points_to_3d
+
 ### Fixed
 - review hash and eq methods
-- ConicalSurface3D: plane_intersections
+- fix pylint.
 
 #### curves.py
 - Ellipse2D/3D: mutualize length method.
@@ -38,6 +48,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ExtrusionSurface3D: fullarcellipse3d_to_2d
 - ExtrusionSurface3D: generalization of the _repair_points_order method to repair the order of parametric points of edges after transformation.
 - ToroidalSurface3D: increases precision of point3d_to_2d.
+- ConicalSurface3D: plane_intersections.
 - Surface3D: repeair_primitives_periodicity. Treat special case on surfaces with singularities.
 
 #### wires.py
@@ -56,6 +67,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### surfaces.py
 - contour3d_to_2d/contour2d_to_3d: Add option to return also a dictionary with the correspondence between the parametric and 3D primitives.
+
+#### display.py
+- refactor DisplayMesh.
 
 ### Changed
 - Edge.split_between_two_points -> trim